HACK: hackly support running vitis locally
