+++ deploy/runtools/firesim_topology_elements.py
@@ -275,45 +275,45 @@ class FireSimServerNode(FireSimNode):
             rootLogger.debug("[localhost] " + str(localcap))
             rootLogger.debug("[localhost] " + str(localcap.stderr))
 
-        # mount rootfs, copy files from it back to local system
-        rfsname = self.get_rootfs_name()
-        if rfsname is not None:
-            is_qcow2 = rfsname.endswith(".qcow2")
-            mountpoint = """/home/centos/sim_slot_{}/mountpoint""".format(simserverindex)
-            with StreamLogger('stdout'), StreamLogger('stderr'):
-                run("""sudo mkdir -p {}""".format(mountpoint))
+        ## mount rootfs, copy files from it back to local system
+        #rfsname = self.get_rootfs_name()
+        #if rfsname is not None:
+        #    is_qcow2 = rfsname.endswith(".qcow2")
+        #    mountpoint = """/home/centos/sim_slot_{}/mountpoint""".format(simserverindex)
+        #    with StreamLogger('stdout'), StreamLogger('stderr'):
+        #        run("""sudo mkdir -p {}""".format(mountpoint))
 
-                if is_qcow2:
-                    rfsname = self.get_host_instance().nbd_tracker.get_nbd_for_imagename(rfsname)
-                else:
-                    rfsname = """/home/centos/sim_slot_{}/{}""".format(simserverindex, rfsname)
+        #        if is_qcow2:
+        #            rfsname = self.get_host_instance().nbd_tracker.get_nbd_for_imagename(rfsname)
+        #        else:
+        #            rfsname = """/home/centos/sim_slot_{}/{}""".format(simserverindex, rfsname)
 
-                run("""sudo mount {blockfile} {mntpt}""".format(blockfile=rfsname, mntpt=mountpoint))
-                with warn_only():
-                    # ignore if this errors. not all rootfses have /etc/sysconfig/nfs
-                    run("""sudo chattr -i {}/etc/sysconfig/nfs""".format(mountpoint))
-                run("""sudo chown -R centos {}""".format(mountpoint))
+        #        run("""sudo mount {blockfile} {mntpt}""".format(blockfile=rfsname, mntpt=mountpoint))
+        #        with warn_only():
+        #            # ignore if this errors. not all rootfses have /etc/sysconfig/nfs
+        #            run("""sudo chattr -i {}/etc/sysconfig/nfs""".format(mountpoint))
+        #        run("""sudo chown -R centos {}""".format(mountpoint))
 
-            ## copy back files from inside the rootfs
-            with warn_only(), StreamLogger('stdout'), StreamLogger('stderr'):
-                for outputfile in jobinfo.outputs:
-                    rsync_cap = rsync_project(remote_dir=mountpoint + outputfile,
-                            local_dir=job_dir,
-                            ssh_opts="-o StrictHostKeyChecking=no",
-                            extra_opts="-L",
-                            upload=False,
-                            capture=True)
-                    rootLogger.debug(rsync_cap)
-                    rootLogger.debug(rsync_cap.stderr)
+        #    ## copy back files from inside the rootfs
+        #    with warn_only(), StreamLogger('stdout'), StreamLogger('stderr'):
+        #        for outputfile in jobinfo.outputs:
+        #            rsync_cap = rsync_project(remote_dir=mountpoint + outputfile,
+        #                    local_dir=job_dir,
+        #                    ssh_opts="-o StrictHostKeyChecking=no",
+        #                    extra_opts="-L",
+        #                    upload=False,
+        #                    capture=True)
+        #            rootLogger.debug(rsync_cap)
+        #            rootLogger.debug(rsync_cap.stderr)
 
-            ## unmount
-            with StreamLogger('stdout'), StreamLogger('stderr'):
-                run("""sudo umount {}""".format(mountpoint))
+        #    ## unmount
+        #    with StreamLogger('stdout'), StreamLogger('stderr'):
+        #        run("""sudo umount {}""".format(mountpoint))
 
-            ## if qcow2, detach .qcow2 image from the device, we're done with it
-            if is_qcow2:
-                with StreamLogger('stdout'), StreamLogger('stderr'):
-                    run("""sudo qemu-nbd -d {devname}""".format(devname=rfsname))
+        #    ## if qcow2, detach .qcow2 image from the device, we're done with it
+        #    if is_qcow2:
+        #        with StreamLogger('stdout'), StreamLogger('stderr'):
+        #            run("""sudo qemu-nbd -d {devname}""".format(devname=rfsname))
 
 
         ## copy output files generated by the simulator that live on the host:
