+++ deploy/runtools/runtime_hw_config.py
@@ -1,0 +1,247 @@
+""" This file manages the overall configuration of the system for running
+simulation tasks. """
+
+from __future__ import print_function
+
+from time import strftime, gmtime
+import ConfigParser
+import pprint
+import logging
+
+from fabric.api import *
+from awstools.awstools import *
+from awstools.afitools import *
+from runtools.firesim_topology_with_passes import FireSimTopologyWithPasses
+from runtools.workload import WorkloadConfig
+from runtools.run_farm import RunFarm
+from util.streamlogger import StreamLogger
+import os
+
+LOCAL_DRIVERS_BASE = "../sim/output/"
+LOCAL_SYSROOT_LIB = "../sim/lib-install/lib/"
+CUSTOM_RUNTIMECONFS_BASE = "../sim/custom-runtime-configs/"
+
+rootLogger = logging.getLogger()
+
+class RuntimeHWConfig:
+    """ A pythonic version of the entires in config_hwdb.ini """
+
+    def __init__(self, name, hwconfig_dict):
+        self.name = name
+
+        self.agfi = None
+        self.xclbin = None
+        self.platform = hwconfig_dict['platform']
+        if self.platform == 'vitis':
+            self.xclbin = hwconfig_dict['xclbin']
+        elif self.platform == 'f1':
+            self.agfi = hwconfig_dict['agfi']
+
+        self.deploytriplet = hwconfig_dict['deploytripletoverride']
+        self.deploytriplet = self.deploytriplet if self.deploytriplet != "None" else None
+        if self.deploytriplet is not None:
+            rootLogger.warning("{} is overriding a deploytriplet in your config_hwdb.ini file. Make sure you understand why!".format(name))
+        self.customruntimeconfig = hwconfig_dict['customruntimeconfig']
+        self.customruntimeconfig = self.customruntimeconfig if self.customruntimeconfig != "None" else None
+        # note whether we've built a copy of the simulation driver for this hwconf
+        self.driver_built = False
+
+    def get_deploytriplet_for_config(self):
+        """ Get the deploytriplet for this configuration. This memoizes the request
+        to the AWS AGFI API."""
+        if self.deploytriplet is not None:
+            return self.deploytriplet
+        rootLogger.debug("Setting deploytriplet by querying the AGFI's description.")
+        self.deploytriplet = get_firesim_tagval_for_agfi(self.agfi,
+                                                         'firesim-deploytriplet')
+    def get_design_name(self):
+        """ Returns the name used to prefix MIDAS-emitted files. (The DESIGN make var) """
+        my_deploytriplet = self.get_deploytriplet_for_config()
+        my_design = my_deploytriplet.split("-")[0]
+        return my_design
+
+    def get_local_driver_binaryname(self):
+        """ Get the name of the driver binary. """
+        return self.get_design_name() + "-" + self.platform
+
+    def get_local_driver_path(self):
+        """ return relative local path of the driver used to run this sim. """
+        my_deploytriplet = self.get_deploytriplet_for_config()
+        drivers_software_base = LOCAL_DRIVERS_BASE + "/" + self.platform + "/" + my_deploytriplet + "/"
+        fpga_driver_local = drivers_software_base + self.get_local_driver_binaryname()
+        return fpga_driver_local
+
+    def get_local_shared_libraries(self):
+        """ Returns a list of path tuples, (A, B), where:
+            A is the local file path on the manager instance to the library
+            B is the destination file path on the runfarm instance relative to the driver """
+
+        return [[LOCAL_SYSROOT_LIB + "/libdwarf.so", "libdwarf.so.1"],
+                [LOCAL_SYSROOT_LIB + "/libelf.so", "libelf.so.1"]]
+
+    def get_local_runtimeconf_binaryname(self):
+        """ Get the name of the runtimeconf file. """
+        return "runtime.conf" if self.customruntimeconfig is None else os.path.basename(self.customruntimeconfig)
+
+    def get_local_runtime_conf_path(self):
+        """ return relative local path of the runtime conf used to run this sim. """
+        my_deploytriplet = self.get_deploytriplet_for_config()
+        drivers_software_base = LOCAL_DRIVERS_BASE + "/" + self.platform + "/" + my_deploytriplet + "/"
+        my_runtimeconfig = self.customruntimeconfig
+        if my_runtimeconfig is None:
+            runtime_conf_local = drivers_software_base + "runtime.conf"
+        else:
+            runtime_conf_local = CUSTOM_RUNTIMECONFS_BASE + my_runtimeconfig
+        return runtime_conf_local
+
+    def get_boot_simulation_command(self, slotid, all_macs,
+                                              all_rootfses, all_linklatencies,
+                                              all_netbws, profile_interval,
+                                              all_bootbinaries, trace_enable,
+                                              trace_select, trace_start, trace_end,
+                                              trace_output_format,
+                                              autocounter_readrate, all_shmemportnames,
+                                              enable_zerooutdram, disable_asserts,
+                                              print_start, print_end,
+                                              enable_print_cycle_prefix):
+        """ return the command used to boot the simulation. this has to have
+        some external params passed to it, because not everything is contained
+        in a runtimehwconfig. TODO: maybe runtimehwconfig should be renamed to
+        pre-built runtime config? It kinda contains a mix of pre-built and
+        runtime parameters currently. """
+
+        # TODO: supernode support
+        tracefile = "+tracefile=TRACEFILE" if trace_enable else ""
+        autocounterfile = "+autocounter-filename=AUTOCOUNTERFILE"
+
+        # this monstrosity boots the simulator, inside screen, inside script
+        # the sed is in there to get rid of newlines in runtime confs
+        driver = self.get_local_driver_binaryname()
+        runtimeconf = self.get_local_runtimeconf_binaryname()
+
+        def array_to_plusargs(valuesarr, plusarg):
+            args = []
+            for index, arg in enumerate(valuesarr):
+                if arg is not None:
+                    args.append("""{}{}={}""".format(plusarg, index, arg))
+            return " ".join(args) + " "
+
+        def array_to_lognames(values, prefix):
+            names = ["{}{}".format(prefix, i) if val is not None else None
+                     for (i, val) in enumerate(values)]
+            return array_to_plusargs(names, "+" + prefix)
+
+        command_macs = array_to_plusargs(all_macs, "+macaddr")
+        command_rootfses = array_to_plusargs(all_rootfses, "+blkdev")
+        command_linklatencies = array_to_plusargs(all_linklatencies, "+linklatency")
+        command_netbws = array_to_plusargs(all_netbws, "+netbw")
+        command_shmemportnames = array_to_plusargs(all_shmemportnames, "+shmemportname")
+        command_dromajo = "+drj_dtb=" + all_bootbinaries[0] + ".dtb" +  " +drj_bin=" + all_bootbinaries[0] + " +drj_rom=" + all_bootbinaries[0] + ".rom"
+
+        command_niclogs = array_to_lognames(all_macs, "niclog")
+        command_blkdev_logs = array_to_lognames(all_rootfses, "blkdev-log")
+
+        command_bootbinaries = array_to_plusargs(all_bootbinaries, "+prog")
+        zero_out_dram = "+zero-out-dram" if (enable_zerooutdram) else ""
+        disable_asserts_arg = "+disable-asserts" if (disable_asserts) else ""
+        print_cycle_prefix = "+print-no-cycle-prefix" if not enable_print_cycle_prefix else ""
+
+        # TODO supernode support
+        dwarf_file_name = "+dwarf-file-name=" + all_bootbinaries[0] + "-dwarf"
+
+        device_index = 0 # unhardcode
+        screen_name = "fsim{}".format(slotid if self.platform == "f1" else device_index)
+        run_device_placement = "+slotid={}".format(slotid) if self.platform == "f1" else "+device_index={}".format(device_index)
+        other = "+binary_file={}".format(self.xclbin) if self.platform == "vitis" else ""
+
+        # TODO: supernode support (tracefile, trace-select.. etc)
+        basecommand = """screen -S {screen_name} -d -m bash -c "script -f -c 'stty intr ^] && LD_LIBRARY_PATH=.:$LD_LIBRARY_PATH ./{driver} +permissive $(sed \':a;N;$!ba;s/\\n/ /g\' {runtimeconf}) {run_device_placement} {other} +profile-interval={profile_interval} {zero_out_dram} {disable_asserts} {command_macs} {command_rootfses} {command_niclogs} {command_blkdev_logs}  {tracefile} +trace-select={trace_select} +trace-start={trace_start} +trace-end={trace_end} +trace-output-format={trace_output_format} {dwarf_file_name} +autocounter-readrate={autocounter_readrate} {autocounterfile} {command_dromajo} {print_cycle_prefix} +print-start={print_start} +print-end={print_end} {command_linklatencies} {command_netbws}  {command_shmemportnames} +permissive-off {command_bootbinaries} && stty intr ^c' uartlog"; sleep 1""".format(
+            screen_name=screen_name,
+            run_device_placement=run_device_placement,
+            other=other,
+            driver=driver,
+            runtimeconf=runtimeconf,
+            command_macs=command_macs,
+            command_rootfses=command_rootfses,
+            command_niclogs=command_niclogs,
+            command_blkdev_logs=command_blkdev_logs,
+            command_linklatencies=command_linklatencies,
+            command_netbws=command_netbws,
+            profile_interval=profile_interval,
+            zero_out_dram=zero_out_dram,
+            disable_asserts=disable_asserts_arg,
+            command_shmemportnames=command_shmemportnames,
+            command_bootbinaries=command_bootbinaries,
+            trace_select=trace_select,
+            trace_start=trace_start,
+            trace_end=trace_end,
+            tracefile=tracefile,
+            trace_output_format=trace_output_format,
+            dwarf_file_name=dwarf_file_name,
+            autocounterfile=autocounterfile,
+            autocounter_readrate=autocounter_readrate,
+            command_dromajo=command_dromajo,
+            print_cycle_prefix=print_cycle_prefix,
+            print_start=print_start,
+            print_end=print_end)
+
+        return basecommand
+
+    def get_kill_simulation_command(self):
+        driver = self.get_local_driver_binaryname()
+        # Note that pkill only works for names <=15 characters
+        return """pkill -SIGKILL {driver}""".format(driver=driver[:15])
+
+    def build_fpga_driver(self):
+        """ Build FPGA driver for running simulation """
+        if self.driver_built:
+            # we already built the driver at some point
+            return
+        # TODO there is a duplicate of this in runtools
+        triplet_pieces = self.get_deploytriplet_for_config().split("-")
+        design = triplet_pieces[0]
+        target_config = triplet_pieces[1]
+        platform_config = triplet_pieces[2]
+        rootLogger.info("Building FPGA software driver for " + str(self.get_deploytriplet_for_config()))
+        with prefix('cd ../'), \
+             prefix('export RISCV={}'.format(os.getenv('RISCV', ""))), \
+             prefix('export PATH={}'.format(os.getenv('PATH', ""))), \
+             prefix('export LD_LIBRARY_PATH={}'.format(os.getenv('LD_LIBRARY_PATH', ""))), \
+             prefix('source ./sourceme-f1-manager.sh'), \
+             prefix('cd sim/'), \
+             StreamLogger('stdout'), \
+             StreamLogger('stderr'):
+            localcap = None
+            with settings(warn_only=True):
+                driverbuildcommand = """make DESIGN={} TARGET_CONFIG={} PLATFORM_CONFIG={} PLATFORM={} driver""".format(design, target_config, platform_config, self.platform)
+                localcap = local(driverbuildcommand, capture=True)
+            rootLogger.debug("[localhost] " + str(localcap))
+            rootLogger.debug("[localhost] " + str(localcap.stderr))
+            if localcap.failed:
+                rootLogger.info("FPGA software driver build failed. Exiting. See log for details.")
+                rootLogger.info("""You can also re-run '{}' in the 'firesim/sim' directory to debug this error.""".format(driverbuildcommand))
+                exit(1)
+
+        self.driver_built = True
+
+
+    def __str__(self):
+        return """RuntimeHWConfig: {}\nDeployTriplet: {}\nAGFI: {}\nCustomRuntimeConf: {}""".format(self.name, self.deploytriplet, self.agfi, str(self.customruntimeconfig))
+
+
+class RuntimeHWDB:
+    """ This class manages the hardware configurations that are available
+    as endpoints on the simulation. """
+
+    def __init__(self, hardwaredbconfigfile):
+        agfidb_configfile = ConfigParser.ConfigParser(allow_no_value=True)
+        agfidb_configfile.read(hardwaredbconfigfile)
+        agfidb_dict = {s:dict(agfidb_configfile.items(s)) for s in agfidb_configfile.sections()}
+
+        self.hwconf_dict = {s: RuntimeHWConfig(s, v) for s, v in agfidb_dict.items()}
+
+    def get_runtimehwconfig_from_name(self, name):
+        return self.hwconf_dict[name]
+
+    def __str__(self):
+        return pprint.pformat(vars(self))
